Wednesday EOB - incorporated functions and moved others to bin
