--- data/combined 2015-2021 exports fft.csv
@@ -1,214 +1,0 @@
-X,t_h,t_d,Year,Month,Week,Day ,Intraday
-615.85,8665.57,361.07,Year,,,,
-205.80,6739.89,280.83,,,,,
-187.07,6065.90,252.75,,,,,
-135.03,3369.94,140.41,,,,,
-132.04,2426.36,101.10,,,,,
-122.41,2637.35,109.89,,,,,
-116.01,11.99,0.50,,,,,Intraday
-110.21,994.41,41.43,,,,,
-104.31,12.01,0.50,,,,,Intraday
-103.22,1956.74,81.53,,,,,
-88.41,1102.89,45.95,,,,,
-85.13,1347.98,56.17,,,,,
-84.53,4666.08,194.42,,,,,
-84.52,24.01,1.00,,,,Day,
-81.89,1189.39,49.56,,,,,
-80.73,1639.43,68.31,,,,,
-80.67,1596.29,66.51,,,,,
-79.22,625.35,26.06,,Month,,,
-78.25,5054.92,210.62,,,,,
-78.06,3192.58,133.02,,,,,
-77.71,2091.69,87.15,,,,,
-74.81,3032.95,126.37,,,,,
-69.67,556.50,23.19,,,,,
-68.30,2757.23,114.88,,,,,
-67.84,673.99,28.08,,Month,,,
-66.67,4332.79,180.53,,,,,
-65.72,7582.38,315.93,,,,,
-65.68,8.00,0.33,,,,,Intraday
-64.84,854.35,35.60,,Month,,,
-63.63,536.81,22.37,,,,,
-63.30,168.03,7.00,,,Week,,
-62.20,418.34,17.43,,,,,
-60.98,2166.39,90.27,,,,,
-60.39,11.97,0.50,,,,,Intraday
-59.17,705.34,29.39,,Month,,,
-58.84,1010.98,42.12,,,,,
-57.87,808.79,33.70,,Month,,,
-57.82,919.08,38.29,,,,,
-57.43,4043.93,168.50,,,Week,,
-57.37,739.74,30.82,,Month,,,
-56.35,485.27,20.22,,,,,
-56.31,3791.19,157.97,,,,,
-55.37,2888.52,120.36,,,,,
-55.16,1838.15,76.59,,,,,
-54.60,1064.19,44.34,,,,,
-54.21,713.64,29.73,,,,,
-52.85,24.00,1.00,,,,Day,
-51.56,541.60,22.57,,,,,
-49.31,3568.18,148.67,,,,,
-48.76,231.52,9.65,,,,,
-48.54,463.05,19.29,,,,,
-48.43,489.19,20.38,,,,,
-48.33,381.50,15.90,,,,,
-47.19,1410.67,58.78,,,,,
-47.04,170.87,7.12,,,,,
-46.31,298.81,12.45,,,,,
-46.29,168.50,7.02,,,,,
-46.08,23.95,1.00,,,,,
-46.01,1123.31,46.80,,,,,
-45.41,164.83,6.87,,,,,
-44.81,2527.46,105.31,,,,,
-44.63,473.90,19.75,,,,,
-43.78,6.00,0.25,,,,,
-43.75,798.14,33.26,,,,,
-43.74,1263.73,52.66,,,,,
-43.63,336.99,14.04,,,,,
-43.17,561.66,23.40,,,,,
-43.16,354.73,14.78,,,,,
-42.77,415.47,17.31,,,,,
-42.77,374.44,15.60,,,,,
-42.62,2333.04,97.21,,,,,
-42.42,23.99,1.00,,,,,
-42.35,730.83,30.45,,,,,
-42.29,452.68,18.86,,,,,
-41.86,551.45,22.98,,,,,
-41.42,697.23,29.05,,,,,
-41.22,23.97,1.00,,,,,
-40.90,879.12,36.63,,,,,
-39.84,324.38,13.52,,,,,
-39.55,777.68,32.40,,,,,
-39.46,1290.62,53.78,,,,,
-39.42,120.12,5.00,,,,,
-39.28,346.62,14.44,,,,,
-39.11,21.01,0.88,,,,,
-38.92,338.88,14.12,,,,,
-38.73,933.22,38.88,,,,,
-38.64,204.24,8.51,,,,,
-38.55,866.56,36.11,,,,,
-38.43,284.78,11.87,,,,,
-37.75,178.41,7.43,,,,,
-37.73,583.26,24.30,,,,,
-37.65,171.35,7.14,,,,,
-37.21,588.92,24.54,,,,,
-36.78,223.83,9.33,,,,,
-36.53,243.61,10.15,,,,,
-36.49,386.36,16.10,,,,,
-36.45,376.76,15.70,,,,,
-36.43,493.16,20.55,,,,,
-36.09,143.74,5.99,,,,,
-35.57,216.64,9.03,,,,,
-35.38,11.99,0.50,,,,,
-35.31,319.26,13.30,,,,,
-35.30,253.80,10.58,,,,,
-35.29,1555.36,64.81,,,,,
-35.25,8.40,0.35,,,,,
-34.95,681.56,28.40,,,,,
-34.76,947.80,39.49,,,,,
-34.71,28.02,1.17,,,,,
-34.62,1784.09,74.34,,,,,
-34.38,618.97,25.79,,,,,
-34.31,312.68,13.03,,,,,
-34.12,594.70,24.78,,,,,
-34.05,819.72,34.15,,,,,
-33.88,326.12,13.59,,,,,
-33.66,294.46,12.27,,,,,
-33.51,497.20,20.72,,,,,
-33.46,72.47,3.02,,,,,
-33.19,203.55,8.48,,,,,
-33.05,226.34,9.43,,,,,
-32.86,27.93,1.16,,,,,
-32.80,518.45,21.60,,,,,
-32.79,361.07,15.04,,,,,
-32.67,722.13,30.09,,,,,
-32.66,272.01,11.33,,,,,
-32.38,24.08,1.00,,,,,
-32.33,365.42,15.23,,,,,
-32.26,56.01,2.33,,,,,
-32.08,2021.97,84.25,,,,,
-31.95,23.96,1.00,,,,,
-31.82,427.18,17.80,,,,,
-31.57,283.45,11.81,,,,,
-31.56,342.71,14.28,,,,,
-31.43,652.25,27.18,,,,,
-31.22,600.58,25.02,,,,,
-31.05,401.72,16.74,,,,,
-31.00,892.04,37.17,,,,,
-30.94,978.37,40.77,,,,,
-30.73,352.67,14.69,,,,,
-30.41,188.38,7.85,,,,,
-30.38,8.00,0.33,,,,,
-30.32,6.00,0.25,,,,,
-30.31,404.39,16.85,,,,,
-30.24,24.15,1.01,,,,,
-30.24,150.52,6.27,,,,,
-30.15,315.93,13.16,,,,,
-30.11,11.94,0.50,,,,,
-30.07,5514.45,229.77,,,,,
-30.03,250.66,10.44,,,,,
-29.96,1144.51,47.69,,,,,
-29.88,245.58,10.23,,,,,
-29.69,659.34,27.47,,,,,
-29.60,527.47,21.98,,,,,
-29.53,218.20,9.09,,,,,
-29.52,223.01,9.29,,,,,
-29.23,12.04,0.50,,,,,
-29.18,638.52,26.60,,,,,
-29.12,157.97,6.58,,,,,
-29.12,174.31,7.26,,,,,
-29.08,149.41,6.23,,,,,
-28.87,307.91,12.83,,,,,
-28.84,1213.18,50.55,,,,,
-28.77,171.84,7.16,,,,,
-28.70,273.24,11.38,,,,,
-28.67,314.30,13.10,,,,,
-28.65,442.77,18.45,,,,,
-28.64,196.94,8.21,,,,,
-28.60,436.40,18.18,,,,,
-28.53,127.97,5.33,,,,,
-28.50,320.95,13.37,,,,,
-28.39,167.57,6.98,,,,,
-28.31,24.03,1.00,,,,,
-28.28,84.84,3.53,,,,,
-28.24,1895.59,78.98,,,,,
-28.10,566.91,23.62,,,,,
-28.05,255.95,10.66,,,,,
-27.98,546.48,22.77,,,,,
-27.81,306.36,12.76,,,,,
-27.80,228.90,9.54,,,,,
-27.73,288.85,12.04,,,,,
-27.71,501.31,20.89,,,,,
-27.61,129.89,5.41,,,,,
-27.57,129.06,5.38,,,,,
-27.53,158.79,6.62,,,,,
-27.46,433.28,18.05,,,,,
-27.44,157.56,6.56,,,,,
-27.23,350.63,14.61,,,,,
-27.22,102.64,4.28,,,,,
-27.17,84.13,3.51,,,,,
-27.08,84.25,3.51,,,,,
-27.06,358.93,14.96,,,,,
-26.99,962.84,40.12,,,,,
-26.99,1684.97,70.21,,,,,
-26.96,139.77,5.82,,,,,
-26.79,133.32,5.55,,,,,
-26.78,33.62,1.40,,,,,
-26.73,75.82,3.16,,,,,
-26.69,148.67,6.19,,,,,
-26.43,88.42,3.68,,,,,
-26.02,160.05,6.67,,,,,
-25.87,125.33,5.22,,,,,
-25.85,143.40,5.98,,,,,
-25.54,169.44,7.06,,,,,
-25.52,187.22,7.80,,,,,
-25.52,24.17,1.01,,,,,
-25.49,24.09,1.00,,,,,
-25.43,33.49,1.40,,,,,
-25.38,191.35,7.97,,,,,
-25.37,58.49,2.44,,,,,
-25.36,767.84,31.99,,,,,
-25.34,111.71,4.65,,,,,
-25.28,168.97,7.04,,,,,
-25.22,175.32,7.30,,,,,
-25.11,173.31,7.22,,,,,
