--- data/combined 2015-2021 load fft.csv
@@ -1,211 +1,0 @@
-X,t_h,t_d,,,,,
-1066.271,8665.571,361.065,Year,,,,
-698.174,168.030,7.001,,,Week,,
-691.572,24.014,1.001,,,,Day,
-338.578,23.947,0.998,,,,Day,
-310.830,84.015,3.501,,,,,
-218.713,24.081,1.003,,,,Day,
-160.271,2888.524,120.355,,,,,
-141.827,1263.729,52.655,,,,,
-134.859,12.007,0.500,,,,,Intraday
-125.392,1102.891,45.954,,,,,
-124.660,2166.393,90.266,,,,,
-122.050,730.831,30.451,,Month,,,
-120.152,84.132,3.505,,,,,
-118.067,23.881,0.995,,,,Day,
-115.122,798.145,33.256,,Month,,,
-113.071,168.497,7.021,,,Week,,
-110.119,2246.630,93.610,,,,,
-106.539,879.116,36.630,,,,,
-103.187,5.999,0.250,,,,,Intraday
-101.284,1237.939,51.581,,,,,
-97.653,3369.944,140.414,,,,,
-97.419,1733.114,72.213,,,,,
-93.904,1479.488,61.645,,,,,
-92.316,1444.262,60.178,,Month,,,
-91.648,5514.455,229.769,,,,,
-91.033,33.625,1.401,,,,,
-90.227,11.990,0.500,,,,,Intraday
-85.192,1784.088,74.337,,,,,
-83.683,463.046,19.294,,,,,
-83.441,24.148,1.006,,,,Day,
-82.809,673.989,28.083,,Month,,,
-82.221,7.997,0.333,,,,,Intraday
-79.863,167.566,6.982,,,,,
-77.931,978.371,40.765,,,,,
-76.860,1083.196,45.133,,,,,
-73.491,3032.950,126.373,,,,,
-73.472,28.018,1.167,,,,,
-69.905,1213.180,50.549,,,,,
-69.666,24.215,1.009,,,,,
-69.659,23.816,0.992,,,,,
-68.363,439.558,18.315,,,,,
-66.582,56.010,2.334,,,,,
-66.431,4332.786,180.533,,,,,
-66.093,1956.742,81.531,,,,,
-65.978,6065.900,252.746,,,,,
-63.262,514.059,21.419,,,,,
-61.276,866.557,36.107,,,,,
-59.817,83.899,3.496,,,,,
-59.244,3568.176,148.674,,,,,
-58.555,1596.289,66.512,,,,,
-58.406,536.805,22.367,,,,,
-54.412,1123.315,46.805,,,,,
-53.069,583.260,24.302,,,,,
-52.873,962.841,40.118,,,,,
-52.636,11.974,0.499,,,,,
-52.483,2091.690,87.154,,,,,
-51.969,459.538,19.147,,,,,
-51.959,2757.227,114.884,,,,,
-51.871,56.062,2.336,,,,,
-51.251,178.935,7.456,,,,,
-51.198,171.839,7.160,,,,,
-51.147,1838.152,76.590,,,,,
-51.117,84.249,3.510,,,,,
-51.044,2527.458,105.311,,,,,
-50.560,24.283,1.012,,,,,
-50.019,748.877,31.203,,,,,
-49.969,381.503,15.896,,,,,
-49.352,23.751,0.990,,,,,
-49.012,84.838,3.535,,,,,
-48.651,2021.967,84.249,,,,,
-48.524,24.004,1.000,,,,,
-48.262,473.898,19.746,,,,,
-47.730,27.928,1.164,,,,,
-47.359,2637.348,109.889,,,,,
-46.707,83.783,3.491,,,,,
-46.336,168.967,7.040,,,,,
-46.250,854.352,35.598,,,,,
-45.605,1045.845,43.577,,,,,
-45.246,42.037,1.752,,,,,
-44.991,230.643,9.610,,,,,
-44.888,8.397,0.350,,,,,
-44.144,842.486,35.104,,,,,
-43.195,522.922,21.788,,,,,
-43.165,23.621,0.984,,,,,
-42.879,258.123,10.755,,,,,
-41.775,588.922,24.538,,,,,
-41.686,493.163,20.548,,,,,
-41.645,190.752,7.948,,,,,
-41.565,2333.038,97.210,,,,,
-41.450,86.532,3.606,,,,,
-41.233,572.255,23.844,,,,,
-41.045,5054.917,210.622,,,,,
-40.958,8.005,0.334,,,,,
-40.909,5.991,0.250,,,,,
-40.391,24.548,1.023,,,,,
-40.311,174.810,7.284,,,,,
-40.121,167.105,6.963,,,,,
-40.100,6.221,0.259,,,,,
-39.447,11.957,0.498,,,,,
-39.047,933.215,38.884,,,,,
-39.009,1410.674,58.778,,,,,
-38.504,2426.360,101.098,,,,,
-38.380,518.453,21.602,,,,,
-37.872,33.606,1.400,,,,,
-37.477,1166.519,48.605,,,,,
-37.241,350.630,14.610,,,,,
-37.037,4043.933,168.497,,,,,
-37.004,11.192,0.466,,,,,
-36.959,436.396,18.183,,,,,
-36.834,551.445,22.977,,,,,
-36.449,224.663,9.361,,,,,
-36.395,4.797,0.200,,,,,
-36.231,1064.193,44.341,,,,,
-36.196,8.012,0.334,,,,,
-36.136,415.473,17.311,,,,,
-35.986,452.679,18.862,,,,,
-35.822,424.189,17.675,,,,,
-35.729,164.834,6.868,,,,,
-35.690,178.409,7.434,,,,,
-35.670,947.797,39.492,,,,,
-35.362,42.008,1.750,,,,,
-35.007,166.646,6.944,,,,,
-34.885,12.912,0.538,,,,,
-34.849,7582.375,315.932,,,,,
-34.813,11.941,0.498,,,,,
-34.708,7.634,0.318,,,,,
-34.590,23.630,0.985,,,,,
-34.529,153.957,6.415,,,,,
-34.402,12.024,0.501,,,,,
-34.057,705.337,29.389,,,,,
-33.990,449.326,18.722,,,,,
-33.913,33.495,1.396,,,,,
-33.546,24.618,1.026,,,,,
-33.329,24.206,1.009,,,,,
-33.270,625.351,26.056,,,,,
-33.057,418.338,17.431,,,,,
-32.727,566.907,23.621,,,,,
-32.605,767.835,31.993,,,,,
-32.336,195.045,8.127,,,,,
-32.322,652.247,27.177,,,,,
-32.248,819.716,34.155,,,,,
-32.073,23.957,0.998,,,,,
-31.843,3192.579,133.024,,,,,
-31.762,83.668,3.486,,,,,
-31.598,787.779,32.824,,,,,
-31.401,777.679,32.403,,,,,
-31.341,348.615,14.526,,,,,
-31.065,24.273,1.011,,,,,
-30.952,23.995,1.000,,,,,
-30.950,1347.978,56.166,,,,,
-30.683,393.890,16.412,,,,,
-30.491,497.205,20.717,,,,,
-30.398,170.390,7.100,,,,,
-30.255,470.225,19.593,,,,,
-29.970,758.238,31.593,,,,,
-29.960,159.210,6.634,,,,,
-29.936,485.272,20.220,,,,,
-29.836,8.839,0.368,,,,,
-29.836,358.929,14.955,,,,,
-29.570,5.995,0.250,,,,,
-29.565,143.742,5.989,,,,,
-29.541,1290.617,53.776,,,,,
-29.526,24.410,1.017,,,,,
-29.483,4.803,0.200,,,,,
-29.387,84.366,3.515,,,,,
-29.200,165.283,6.887,,,,,
-29.197,322.654,13.444,,,,,
-29.174,169.439,7.060,,,,,
-29.130,388.840,16.202,,,,,
-29.127,501.314,20.888,,,,,
-28.676,489.185,20.383,,,,,
-28.645,218.986,9.124,,,,,
-28.535,202.197,8.425,,,,,
-28.469,164.388,6.849,,,,,
-28.375,28.005,1.167,,,,,
-28.283,697.230,29.051,,,,,
-28.283,10.493,0.437,,,,,
-28.274,681.562,28.398,,,,,
-27.918,442.766,18.449,,,,,
-27.801,139.126,5.797,,,,,
-27.694,4666.077,194.420,,,,,
-27.635,1144.509,47.688,,,,,
-27.560,236.027,9.834,,,,,
-27.445,23.312,0.971,,,,,
-27.408,631.865,26.328,,,,,
-27.399,159.629,6.651,,,,,
-27.223,83.209,3.467,,,,,
-27.073,1028.119,42.838,,,,,
-26.754,340.781,14.199,,,,,
-26.678,24.033,1.001,,,,,
-26.530,5.792,0.241,,,,,
-26.496,1516.475,63.186,,,,,
-26.349,267.220,11.134,,,,,
-26.326,379.119,15.797,,,,,
-26.171,204.239,8.510,,,,,
-26.097,541.598,22.567,,,,,
-26.039,186.643,7.777,,,,,
-25.961,23.985,0.999,,,,,
-25.898,23.686,0.987,,,,,
-25.888,28.109,1.171,,,,,
-25.750,689.307,28.721,,,,,
-25.592,24.351,1.015,,,,,
-25.574,7.990,0.333,,,,,
-25.482,151.269,6.303,,,,,
-25.323,33.756,1.406,,,,,
-25.323,3791.188,157.966,,,,,
-25.272,23.250,0.969,,,,,
-25.154,363.228,15.134,,,,,
-25.090,477.630,19.901,,,,,
-25.076,24.759,1.032,,,,,
