--- data/combined 2015-2021 prices fft.csv
@@ -1,135 +1,0 @@
-X,t_h,t_d
-288.8032741597468,8665.57142857143,361.0654761904762
-187.4143932464734,168.03047091412742,7.001269621421976
-168.7272444397201,24.013855898653997,1.0005773291105833
-142.7737508020431,7582.375,315.9322916666667
-141.80959370342103,11.99031429136193,0.49959642880674704
-137.35415567387767,6739.888888888889,280.8287037037037
-122.5950758935119,6065.9,252.74583333333334
-119.93408184493572,4666.076923076923,194.4198717948718
-110.65635380249019,5514.454545454545,229.76893939393938
-105.9596363185424,24.080587534736004,1.0033578139473336
-97.76145981445518,84.01523545706371,3.500634810710988
-96.66931323506365,5054.916666666667,210.62152777777777
-96.13199446843475,12.006927949326998,0.5002886645552916
-88.40951830182573,4043.9333333333334,168.49722222222223
-86.85344038534663,3192.578947368421,133.02412280701753
-83.84316548949785,3791.1875,157.96614583333334
-82.89868790674012,3568.1764705882356,148.67401960784315
-73.34107484315781,1444.2619047619048,60.17757936507937
-67.6924601553308,1733.1142857142856,72.21309523809524
-66.44856575958792,2757.2272727272725,114.88446969696969
-63.15639348819072,2637.3478260869565,109.88949275362319
-62.32576225429043,4332.785714285715,180.5327380952381
-61.73582380168461,3369.9444444444443,140.41435185185185
-61.055762760767024,1237.938775510204,51.58078231292517
-60.55448426492988,23.881496062992124,0.9950623359580051
-57.876678591539466,1639.4324324324323,68.30968468468468
-56.97629943777747,1784.0882352941178,74.33700980392157
-56.12988722686256,1045.8448275862067,43.57686781609195
-56.03331085968425,1083.1964285714287,45.133184523809526
-53.28139912370422,2426.36,101.09833333333333
-53.22158792687458,1684.9722222222222,70.20717592592592
-52.4491031817427,167.56629834254142,6.981929097605893
-52.32802459650359,1516.475,63.186458333333334
-51.29043545584239,1166.5192307692307,48.60496794871795
-50.84414076330028,2888.5238095238096,120.35515873015873
-50.41042846104272,2527.4583333333335,105.31076388888889
-49.704311736466515,1596.2894736842104,66.51206140350877
-48.55637900688679,962.8412698412696,40.11838624338624
-48.47558745511284,536.8053097345132,22.36688790560472
-48.1061945425184,2333.0384615384614,97.2099358974359
-48.01960948069637,2021.9666666666667,84.24861111111112
-47.65094863729235,842.4861111111111,35.10358796296296
-47.627851172411624,1213.18,50.549166666666665
-47.47618426637638,1010.9833333333333,42.12430555555556
-46.88929415227237,1064.1929824561405,44.341374269005854
-46.54302171724246,11.973746545598104,0.4989061060665877
-45.65128553129063,24.14769108280255,1.006153795116773
-45.00923508250387,1347.9777777777779,56.16574074074074
-44.33887520961077,854.3521126760563,35.59800469483568
-43.809256763112316,600.5841584158416,25.0243399339934
-43.10510978424327,572.2547169811321,23.843946540880502
-42.79053262989336,11.957224521979107,0.4982176884157961
-42.026020475515146,167.1046831955923,6.962695133149678
-41.99305630907059,28.01801385681293,1.1674172440338721
-41.970093629491785,819.7162162162161,34.15484234234234
-41.5559557490536,5.999307684699832,0.24997115352915966
-41.50443586094514,12.040293767368002,0.5016789069736668
-41.38156268596014,933.2153846153846,38.883974358974356
-41.13463032128348,546.4774774774775,22.769894894894897
-41.045947138693855,165.283378746594,6.886807447774749
-40.54970999739668,730.8313253012049,30.451305220883537
-40.480897043827184,23.947493091196208,0.9978122121331754
-40.37934126125444,919.0757575757576,38.29482323232323
-40.21353322518287,1318.6739130434783,54.944746376811594
-39.98719876704647,866.5571428571428,36.10654761904762
-39.76101548964991,1555.3589743589744,64.80662393162393
-39.59374859632485,1144.5094339622642,47.687893081761004
-39.335475698652175,830.945205479452,34.62271689497717
-39.093944815429225,1838.1515151515152,76.58964646464646
-38.80632352105635,166.6456043956044,6.9435668498168495
-38.70402870693632,947.796875,39.491536458333336
-38.62156953534603,7.997231377719183,0.3332179740716326
-38.040173416135694,21.011084170419117,0.8754618404341299
-37.967504831336434,12.023587710604557,0.5009828212751899
-37.483075799372884,1028.1186440677966,42.83827683615819
-37.27357136857114,748.8765432098766,31.203189300411523
-37.04439262941341,1895.59375,78.98307291666667
-35.13492320834637,166.18904109589042,6.924543378995434
-34.78324971524917,879.1159420289855,36.62983091787439
-34.74379955209117,33.62472283813747,1.401030118255728
-34.719521879892326,1189.3921568627452,49.55800653594771
-34.07996785263926,1410.674418604651,58.77810077519379
-33.38878616801697,383.9177215189874,15.996571729957806
-33.09854132321024,1378.6136363636363,57.442234848484844
-32.99366634796748,218.98555956678697,9.124398315282791
-32.86960356744108,1479.4878048780488,61.645325203252035
-32.86313690231042,3032.95,126.37291666666667
-32.86081868860617,280.8287037037037,11.70119598765432
-32.668102252914,463.0458015267176,19.293575063613233
-32.534157095768826,514.0593220338983,21.419138418079093
-32.09132737476828,83.78314917127071,3.4909645488029466
-31.930842230423515,11.940748031496062,0.49753116797900254
-31.91734131350462,56.01015697137581,2.333756540473992
-31.905893751984202,618.969387755102,25.790391156462587
-31.64817962614874,583.2596153846154,24.302483974358974
-30.851208763393625,24.21516966067864,1.0089654025282766
-30.774822425056907,23.815861798193954,0.9923275749247481
-30.65047938800285,493.1626016260163,20.548441734417345
-30.597702921809034,83.89903181189489,3.495792992162287
-30.13672434617239,561.6574074074074,23.40239197530864
-30.109989076822224,293.0386473429952,12.209943639291467
-29.479426283359196,758.2375,31.593229166666667
-29.463168888871117,577.7047619047619,24.071031746031746
-29.20140536094927,905.358208955224,37.72325870646766
-29.043703569755852,56.06192236598891,2.3359134319162047
-28.629361267678604,212.0944055944056,8.8372668997669
-28.356572701266927,11.995056357524225,0.4997940148968427
-28.18145097595816,588.9223300970873,24.538430420711972
-28.148026447074013,994.4098360655737,41.43374316939891
-28.04012422533574,2246.6296296296296,93.60956790123456
-28.006192308072116,606.59,25.274583333333332
-27.915285650361312,892.0441176470589,37.16850490196079
-27.774773265016474,612.7171717171717,25.529882154882156
-27.59869925992812,228.04135338345867,9.50172305764411
-27.456174667940246,1290.617021276596,53.775709219858165
-27.30162026579951,165.73497267759564,6.905623861566485
-27.120637399985593,258.12340425531914,10.75514184397163
-27.065180582153165,363.22754491017963,15.134481037924152
-27.01602665794214,11.98557597312784,0.4993989988803267
-26.955787984822333,11.98794466403162,0.49949769433465085
-26.919342340666834,168.96657381615597,7.040273909006499
-26.67211043394682,84.13176144244106,3.5054900601017107
-26.66446502765887,798.1447368421052,33.25603070175438
-26.51358242668789,216.6392857142857,9.026636904761904
-26.424786083244687,421.24305555555554,17.55179398148148
-26.390656894233402,594.6960784313726,24.779003267973856
-26.25999780888981,290.23444976076553,12.09310207336523
-26.208423645201,284.78403755868544,11.866001564945226
-26.13827336172226,309.484693877551,12.895195578231293
-25.72160845244369,224.66296296296298,9.360956790123458
-25.27382970889309,7.9898577449947314,0.3329107393747805
-25.264292754877815,439.55797101449275,18.314915458937197
-25.195247379827258,689.3068181818181,28.721117424242422
-25.146488233635285,241.66932270916337,10.069555112881806
