+++ requirements.txt
@@ -1,7 +1,9 @@
+matplotlib>=3.2.0
+numpy>=1.22.0
+openpyxl>=3.0.0
+pandas>=1.0.0
 python>=3.8.0
-matplotlib==3.2.1
-numpy==1.22.3
-openpyxl==3.0.9
-pandas==1.0.3
-xlrd==2.0.1
-xlsxwriter==3.0.3
+scipy>=1.8.0
+seaborn>=0.11.0
+xlrd>=2.0.0
+xlsxwriter>=3.0.0
