+++ to_bin/alternative_timeslices.md
@@ -1,0 +1,13 @@
+# Alternative timeslices
+
+_14 April 2022_
+
+* Hourly (24) and monthly (12) timeslices
+* Weekday-weekend (2),
+* Two night, bihourly day slices (2 + 6)\
+
+Progress so far:
+
+* Set up new set_timeslice functions for weekday-weekend, 2+6,
+* started with changing load_duration graphs to accommodate other time slices,
+*
