+++ to_bin/temp-plotDistribution.py
@@ -9,8 +9,8 @@ os.chdir(directory)
 files = os.listdir(directory)
 
 for file in files:
-    file = "C:/Users/czpkersten/Documents/timeslices-output/Output load/Statistics load - season weekday - daynite.csv"
-    #path = directory + file
+    path = directory + file
+    
     df = pd.read_csv(file)
 
     sns.scatterplot(x='Mean', y='Timeslice', data=df, color='b', label='Mean')
@@ -26,10 +26,8 @@ for file in files:
         # Statistical information
         sns.scatterplot(x=column, y='Timeslice', data=df, color='r', marker='x')
 
-    plt.title(file.replace('.csv',''))
-    plt.title('Season - weekday - daynite')
+    plt.title(file.replace('.csv','').lower())
     plt.grid()
     plt.tight_layout()
     plt.savefig(file.replace('.csv', '.png'), dpi=300, format='png')
     plt.show()
-    quit()
